完成商品分类的增删改。。 成功的提示框，没有改好

+++ annie-shop-platform/src/main/java/com/annie/dao/ProductTypeMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface ProductTypeMapper {
     int insertProductType(ProductType record);
 
-    ProductType selectByPrimaryKey(Integer id);
+    ProductType selectByPrimaryKey(Long id);
 
     int updateProductType(ProductType record);
